requirements.txt
addict==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
cycler==0.11.0
Cython==3.0.12
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
iopath==0.1.10
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.5.3
mdurl==0.1.2
mmcv-full==1.3.17
mmdet==2.17.0
mmengine==0.10.7
mmsegmentation==0.19.0
numpy==1.21.6
opencv-python==4.11.0.86
packaging==24.0
Pillow==9.5.0
pip==22.1.2
platformdirs==4.0.0
portalocker==2.7.0
prettytable==3.7.0
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.8.1
scipy==1.7.3
setuptools==65.6.3
six==1.17.0
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
timm==0.6.12
tomli==2.0.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
wheel==0.37.1
yacs==0.1.8
yapf==0.43.0
zipp==3.15.0
